% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asda_theme.R
\name{asda_theme}
\alias{asda_theme}
\title{ASDa ggplot2 Theme}
\usage{
asda_theme(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{the base font size}

\item{base_family}{the default font family}
}
\value{
a ggplot2 theme
}
\description{
A ggplot2 theme based on the black/white theme
from ggplot2 with different margins and grid lines.
}
\details{
Gray, dotted grid lines (color "gray30")
and smaller margins around the plot.
}
