% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colorfill.R
\name{scale_colorfill_brewer}
\alias{scale_colorfill_brewer}
\title{Identical Scales for Color and Fill}
\usage{
scale_colorfill_brewer(..., type = "seq", palette = 1, direction = 1)
}
\arguments{
\item{..., type, palette, direction}{see arguments to \code{\link[ggplot2]{scale_color_brewer}}}
}
\value{
ggplot2 scales for color and fill
}
\description{
Define the same scale for the color and the fill
in ggplot2 plots.
}
