% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colorfill.R
\name{scale_colorfill_gradient}
\alias{scale_colorfill_gradient}
\title{Identical Scales for Color and Fill}
\usage{
scale_colorfill_gradient(..., low = "#132B43", high = "#56B1F7",
  space = "Lab", guide = "colourbar")
}
\arguments{
\item{..., low, high, space, guide}{see arguments to \code{\link[ggplot2]{scale_color_gradient}}}
}
\value{
ggplot2 scales for color and fill
}
\description{
Define the same scale for the color and the fill
in ggplot2 plots.
}
