% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colorfill.R
\name{scale_colorfill_hue}
\alias{scale_colorfill_hue}
\title{Identical Scales for Color and Fill}
\usage{
scale_colorfill_hue(..., h = c(0, 360) + 15, c = 100, l = 65,
  h.start = 0, direction = 1)
}
\arguments{
\item{..., h, c, l, h.start, direction}{see arguments to \code{\link[ggplot2]{scale_color_hue}}}
}
\value{
ggplot2 scales for color and fill
}
\description{
Define the same scale for the color and the fill
in ggplot2 plots.
}
